Moved sns client
