+++ client/sns/message.go
