+++ client/sns/message_test.go
