+++ client/sns/publisher.go
