+++ client/sns/publisher_integration_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sns"
 	"github.com/aws/aws-sdk-go/service/sns/snsiface"
@@ -36,8 +35,7 @@ func createAPI(t *testing.T) snsiface.SNSAPI {
 	sess, err := session.NewSession(
 		aws.NewConfig().
 			WithEndpoint(testSnsEndpoint).
-			WithRegion(testSnsRegion).
-			WithCredentials(credentials.NewStaticCredentials("test", "test", "")),
+			WithRegion(testSnsRegion),
 	)
 	require.NoError(t, err)
