+++ client/sns/publisher_test.go
